# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export type SignupInputRole = typeof SignupInputRole[keyof typeof SignupInputRole];


export const SignupInputRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface SignupInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: SignupInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Project {
  id: number;
  title: string;
  description: string;
  ownerId: number;
  owner?: User;
  members: User[];
  /** @nullable */
  taskCount?: number | null;
  /** @nullable */
  completedTaskCount?: number | null;
  createdAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  description: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
}

export interface AddMemberInput {
  userId: number;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Task {
  id: number;
  title: string;
  description: string;
  projectId: number;
  project?: Project;
  /** @nullable */
  assignedToId?: number | null;
  assignedTo?: User | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  isOverdue?: boolean | null;
  createdById: number;
  createdBy?: User;
  createdAt: string;
  updatedAt: string;
}

export type TaskInputStatus = typeof TaskInputStatus[keyof typeof TaskInputStatus];


export const TaskInputStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description: string;
  /** @nullable */
  assignedToId?: number | null;
  status: TaskInputStatus;
  priority: TaskInputPriority;
  /** @nullable */
  dueDate?: string | null;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  /** @nullable */
  assignedToId?: number | null;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  /** @nullable */
  dueDate?: string | null;
}

export interface Comment {
  id: number;
  body: string;
  taskId: number;
  authorId: number;
  author?: User;
  createdAt: string;
}

export interface CommentInput {
  /** @minLength 1 */
  body: string;
}

export type DashboardStatsTasksByStatus = {
  todo: number;
  in_progress: number;
  completed: number;
};

export interface DashboardStats {
  totalProjects: number;
  totalUsers: number;
  totalTasks: number;
  completedTasks: number;
  overdueTasks: number;
  completionPercentage: number;
  tasksByStatus: DashboardStatsTasksByStatus;
  recentActivity: Task[];
  assignedProjects?: Project[];
  assignedTasks?: Task[];
}