# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface Comment {
  id: number;
  body: string;
  taskId: number;
  authorId: number;
  author?: User;
  createdAt: Date;
}