# lib/api-zod/src/generated/types/commentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CommentInput {
  /** @minLength 1 */
  body: string;
}