# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  description: string;
}